services and mappings tests added

+++ FootballLeagueWebAPI/Services/LeagueInputService.cs
@@ -5,12 +5,13 @@ using System;
 
 namespace FootballLeagueWebAPI.Services
 {
-    public class LeagueInputService
+    public class LeagueInputService : ILeagueInputService
     {
         private readonly TeamRepository _teamRepository;
         private readonly PlayerRepositiory _playerRepositiory;
         private readonly MatchRepository _matchRepository;
         private (int Team, int Player, int Match) StartingId = (0, 0, 0);
+
         public LeagueInputService(TeamRepository teamRepository, PlayerRepositiory playerRepositiory, MatchRepository matchRepositiory)
         {
             _teamRepository = teamRepository;
@@ -38,6 +39,11 @@ namespace FootballLeagueWebAPI.Services
 
         public void RemoveTeamById(int teamId)
         {
+            if(!_teamRepository.IdExists(teamId))
+            {
+                throw new ArgumentException();
+            }
+
             _teamRepository.RemoveById(teamId);
         }
 
@@ -59,6 +65,12 @@ namespace FootballLeagueWebAPI.Services
 
         public void TransferPlayer(int playerId, int newTeamId)
         {
+            if(!_playerRepositiory.IdExists(playerId) || 
+                !_teamRepository.IdExists(newTeamId))
+            {
+                throw new ArgumentException();
+            }
+
             _teamRepository.RemovePlayer(playerId);
 
             _teamRepository.AddPlayer(newTeamId, _playerRepositiory.GetById(playerId));
@@ -66,11 +78,24 @@ namespace FootballLeagueWebAPI.Services
 
         public void RemovePlayerById(int playerId)
         {
+            if(!_playerRepositiory.IdExists(playerId))
+            {
+                throw new ArgumentException();
+            }
+
             _playerRepositiory.RemoveById(playerId);
         }
 
         public void PlayMatch(MatchRequest matchRequest)
         {
+            if(matchRequest.GuestTeamGoals<0 || matchRequest.HomeTeamGoals<0 || 
+                !_teamRepository.IdExists(matchRequest.HomeTeamId) || 
+                !_teamRepository.IdExists(matchRequest.GuestTeamId) ||
+                matchRequest.GuestTeamId == matchRequest.HomeTeamId)
+            {
+                throw new ArgumentException();
+            }
+
             Match match = new Match
             {
                 Id = StartingId.Match++,
